Add files via upload
This project aims to target customer to offer bank loan offer. Three types of classification models are built, models include decision tree, KNN, and logistic regression. Due to the imbalance of the two classes, we also built other 3 models on the resampled data. The best performance occurs when using the decision tree model after the resampling. The accuracy is 0.98, with 0.94 AUC score.